bachelor/myapp/pyproject.toml
[project]
name = "myapp"
version = "0.1.0"
description = ""
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.12,<3.14"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydantic (>=2.11.9,<3.0.0)",
    "supabase (>=2.18.1,<3.0.0)",
    "b2sdk (>=2.10.0,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "torch (>=2.8.0,<3.0.0)",
    "pinecone (>=7.3.0,<8.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "transformers (>=4.56.2,<5.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "trimesh (>=4.10.1,<5.0.0)",
    "pyrender (>=0.1.45,<0.2.0)",
    "opencv-python-headless (>=4.12.0.88,<5.0.0.0)",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
